src/azure_functions/requirements.txt
azure-functions
azure-storage-blob
azure-storage-queue
httpx
tiktoken
python-dotenv
